Ensure that only one redis connection is used per request
When testing the code locally, everything worked as expected. But when
running it in production, it quickly hang.

The reason might be a connection pool exhaustion. Multiple requests
request a connection, but none are available because they are hold by
other requests. The ones already holding a connection might request
another one and we run into a deadlock.

I'm not sure if this is really what is happening, but it seems to be a
possibility.