Cache infrequent changing Redis data
Data like software and chart configuration hardly ever changes (at most,
more are added). Since we read this data on every data submission,
caching this data in-memory results in a **huge** performance increase.

The downside is that this increases memory usage and does not scale well
horizontally. But that is an acceptable compromise.